requirements.txt
python-binance==1.0.19
pandas==2.2.2
pandas-ta>=0.3.14b0
anthropic==0.40.0
openai==1.51.0
google-generativeai==0.8.3
python-telegram-bot==21.5
apscheduler==3.10.4
aiosqlite==0.20.0
pandas-datareader==0.10.0
newsapi-python==0.2.7
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
requests==2.32.3
psutil>=5.9.0
